backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
psycopg2-binary>=2.9.6
python-decouple>=3.8
Pillow>=10.0.0
gunicorn>=21.2.0
